Tests pass

+++ ftc/test_selenium.py
@@ -1,6 +1,7 @@
 from django.test import Client, tag, TestCase
 from pathlib import Path
 from selenium import webdriver
+from selenium.common import exceptions
 from selenium.webdriver.common.alert import Alert
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
@@ -48,9 +49,26 @@ class DockerCompose:
         return self
 
 
+def maybe_click(elt):
+    if elt is None:
+        return False
+    elt.click()
+    return True
+
 class BasePage:
     def __init__(self, driver):
         self.driver = driver
+
+    def check(self):
+        return True
+
+    def wait(self):
+        """
+        Wait until check() passes
+        """
+        WebDriverWait(self.driver, timeout=3).until(
+            lambda d: self.check()
+        )
 
     def fill_form(self, fields):
         for k,v in fields.items():
@@ -86,6 +104,35 @@ class BasePage:
             r['y'] + (r['height'] - wr['height'])/2
         )
         self.driver.execute_script(js)
+
+    def find_by_id(self, id):
+        return WebDriverWait(self.driver, timeout=2).until(
+            lambda d: d.find_element(By.ID, id)
+        )
+
+    def find_by_css(self, css):
+        return WebDriverWait(self.driver, timeout=2).until(
+            lambda d: d.find_element(By.CSS_SELECTOR, css)
+        )
+
+    def click_element(self, locator, css):
+        WebDriverWait(
+            self.driver,
+            timeout=2,
+            ignored_exceptions=[
+                exceptions.StaleElementReferenceException,
+                exceptions.ElementNotInteractableException
+            ]
+        ).until(
+            lambda d: maybe_click(d.find_element(locator, css))
+        )
+
+    def click_by_id(self, id):
+        self.click_element(By.ID, id)
+
+    def click_by_css(self, css):
+        self.click_element(By.CSS_SELECTOR, css)
+
 
 class WebMail(BasePage):
     def go(self):
@@ -216,7 +263,7 @@ class ProfilePage(BasePage):
         return self
 
     def get_counting_link(self):
-        return self.driver.find_element(By.ID, 'start-counting-link')
+        return self.find_by_id('start-counting-link')
 
     def go_start_counting(self):
         self.get_counting_link().click()
@@ -245,7 +292,7 @@ class ProfilePage(BasePage):
 
 class CountingPage(BasePage):
     def check(self):
-        self.driver.find_element(By.ID, "btn-tracknum")
+        self.find_by_id('btn-tracknum')
         return self
 
     def go(self):
@@ -277,7 +324,7 @@ class CountingPage(BasePage):
         return self
 
     def delete_from(self, minx, maxx, miny, maxy):
-        self.driver.find_element(By.ID, "ftc-btn-select").click()
+        self.click_by_id('ftc-btn-select')
         mp = self.driver.find_element(By.ID, "map")
         lil = mp.find_element(By.CSS_SELECTOR,  'img.leaflet-image-layer')
         w = lil.size["width"]
@@ -344,9 +391,10 @@ class CountingPage(BasePage):
 
 class NavBar(BasePage):
     def get_dropdown(self):
-        return self.driver.find_element(By.CSS_SELECTOR,
-            "#account-dropdown a"
-        )
+        return self.find_by_css('#account-dropdown a')
+
+    def click_dropdown(self):
+        return self.click_by_css('#account-dropdown a')
 
     def check(self):
         self.get_dropdown()
@@ -357,7 +405,7 @@ class NavBar(BasePage):
             def __init__(self, nav_bar):
                 self.nav = nav_bar
             def __call__(self, driver):
-                self.nav.get_dropdown().click()
+                self.nav.click_dropdown()
                 logout = driver.find_element(By.CSS_SELECTOR,
                     'a[href="/accounts/logout/"]'
                 )
@@ -366,14 +414,14 @@ class NavBar(BasePage):
         return HomePage(self.driver)
 
     def go_manage_projects(self):
-        self.get_dropdown().click()
+        self.click_dropdown()
         self.driver.find_element(By.CSS_SELECTOR,
             'a[href="/ftc/report/"]'
         ).click()
         return ReportPage(self.driver)
 
     def go_edit_projects(self):
-        self.get_dropdown().click()
+        self.click_dropdown()
         self.driver.find_element(By.ID, 'projects-link').click()
         return ProjectsPage(self.driver)
 
@@ -388,11 +436,11 @@ class ReportPage(BasePage):
         return self
 
     def select_tree_node(self, name):
-        self.driver.find_element(By.XPATH,
+        self.click_element(By.XPATH,
             "//span[@class='fancytree-title' and text()='"
             + name
             + "']//preceding-sibling::span[@class='fancytree-checkbox']"
-        ).click()
+        )
         return self
 
     def result_now(self, grain_number):
@@ -464,15 +512,55 @@ class GrainCreatePage(BasePage):
         return GrainPage(self.driver)
 
 
+def find_best(a, f):
+    """
+    Returns the element of `a` for which `f` returns the lowest value
+    """
+    dist = None
+    best = None
+    for v in a:
+        d = f(v)
+        if dist is None or d < dist:
+            dist = d
+            best = v
+    return best
+
+
+def sum_squares(*a):
+    """
+    Returns the sum of the squares of the elements of `a`
+    """
+    total = 0
+    for x in a:
+        total += x * x
+    return total
+
+
+def pin_x(elt):
+    return elt.rect['x'] + elt.rect['width'] / 2
+
+def pin_y(elt):
+    return elt.rect['y'] + elt.rect['height']
+
+
 class GrainPage(BasePage):
-    def get_image_layer(self):
-        return self.driver.find_elements(By.CSS_SELECTOR, 'img.leaflet-image-layer')
+    def get_image_width(self):
+        return WebDriverWait(
+            self.driver,
+            timeout=2,
+            ignored_exceptions=exceptions.StaleElementReferenceException
+        ).until(
+            lambda d: d.find_element(
+                By.CSS_SELECTOR,
+                'img.leaflet-image-layer'
+            ).rect['width']
+        )
 
     def edit(self):
         zoom_outs = self.driver.find_elements(By.CLASS_NAME, 'leaflet-control-zoom-out')
         assert len(zoom_outs) == 1
         zoom_out = zoom_outs[0]
-        while (600 < self.get_image_layer()[0].rect['width']
+        while (600 < self.get_image_width()
                 and not self.element_is_disabled(zoom_out)):
             zoom_out.click()
         self.driver.find_element(By.ID, 'edit').click()
@@ -482,24 +570,34 @@ class GrainPage(BasePage):
         self.driver.find_element(By.ID, 'save').click()
         return self
 
+    def do_partial_drag(self, dragee, dest_x, dest_y):
+        drag_x_by = dest_x - pin_x(dragee)
+        drag_y_by = dest_y - pin_y(dragee)
+        if -2 < drag_x_by and drag_x_by < 2 and -2 < drag_y_by and drag_y_by < 2:
+            return False
+        max_distance = 199
+        drag_x_by = min(max(drag_x_by, -max_distance), max_distance)
+        drag_y_by = min(max(drag_y_by, -max_distance), max_distance)
+        ActionChains(self.driver).move_by_offset(drag_x_by, drag_y_by).perform()
+        return True
+
     def do_drag(self, dragee, dest_x, dest_y):
         ActionChains(self.driver).click_and_hold(dragee).perform()
-        for n in range(3):
-            drag_x_by = dest_x - (
-                dragee.rect['x'] + dragee.rect['width']/2)
-            drag_y_by = dest_y - (
-                dragee.rect['y'] + dragee.rect['height'])
-            ActionChains(self.driver).move_by_offset(
-                drag_x_by, drag_y_by
-            ).perform()
+        while self.do_partial_drag(dragee, dest_x, dest_y):
+            pass
         ActionChains(self.driver).release().perform()
 
-    def drag_marker(self, marker_index, to_x, to_y):
-        imgs = self.driver.find_elements(By.CSS_SELECTOR, 'img.leaflet-image-layer')
-        rect = imgs[0].rect
-        markers = self.driver.find_elements(By.CSS_SELECTOR,
-            'img.leaflet-marker-icon')
-        marker = markers[marker_index]
+    def drag_marker(self, from_x_approx, from_y_approx, to_x, to_y):
+        img = self.driver.find_element(By.CSS_SELECTOR, 'img.leaflet-image-layer')
+        rect = img.rect
+        markers = self.driver.find_elements(
+            By.CSS_SELECTOR,
+            'img.region-vertex-marker'
+        )
+        marker = find_best(markers, lambda m: sum_squares(
+            from_x_approx * rect['width'] + rect['x'] - pin_x(m),
+            from_y_approx * rect['height'] + rect['y'] - pin_y(m)
+        ))
         self.do_drag(
             marker,
             to_x * rect['width'] + rect['x'],
@@ -561,10 +659,10 @@ class WebUploader:
         sample_page = project_page.create_sample().create('s1', 'T', 1, 1, False)
         grain_page = sample_page.create_grain().create(files)
         grain_page.edit()
-        grain_page.drag_marker(0, 0.01, 0.01)
-        grain_page.drag_marker(1, 0.01, 0.99)
-        grain_page.drag_marker(3, 0.99, 0.01)
-        grain_page.drag_marker(2, 0.01, 0.99) # delete by dragging onto 3
+        grain_page.drag_marker(0, 0, 0.01, 0.01)
+        grain_page.drag_marker(0, 1, 0.01, 0.99)
+        grain_page.drag_marker(1, 0, 0.99, 0.01)
+        grain_page.drag_marker(1, 1, 0.01, 0.99) # delete by dragging onto 2nd
         grain_page.save()
 
     def get_index(self, file_url):
@@ -577,17 +675,20 @@ class WebUploader:
 class DjangoTests(TestCase):
     def setUp(self):
         self.dc = DockerCompose("./docker-compose-test.yml").down().up().init()
-        self.tmp = tempfile.mkdtemp(prefix='tmp', dir=Path.home())
-        service = webdriver.firefox.service.Service(service_args=[
-            "--profile-root",
-            self.tmp
-        ])
-        self.driver = webdriver.Firefox(service=service)
+        self.tmp = None
+        #self.tmp = tempfile.mkdtemp(prefix='tmp', dir=Path.home())
+        #service = webdriver.firefox.service.Service(service_args=[
+        #    "--profile-root",
+        #    self.tmp
+        #])
+        #self.driver = webdriver.Firefox(service=service)
+        self.driver = webdriver.Chrome()
 
     def tearDown(self):
         self.driver.close()
         self.dc.down()
-        os.rmdir(self.tmp)
+        if self.tmp is not None:
+            os.rmdir(self.tmp)
 
     def test_onboard(self):
         # Upload Z-Stack images
@@ -629,7 +730,7 @@ class DjangoTests(TestCase):
 
         # do the same thing with John
         profile = SignInPage(self.driver).go().sign_in(project_user)
-        profile.check_cannot_count().go_tutorial().get_to_end_and_finish()
+        profile.check_cannot_count().go_tutorial().get_to_end_and_finish().wait()
         profile.go().check_can_count()
         navbar.logout()
